requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.12.1
arrow==1.4.0
asttokens==3.0.1
attrs==25.4.0
binaryornot==0.4.4
black==25.12.0
certifi==2026.1.4
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cookiecutter==2.6.0
cycler==0.12.1
decorator==5.2.1
executing==2.2.1
filelock==3.20.2
flake8==7.3.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.4
idna==3.11
iniconfig==2.3.0
ipdb==0.13.13
ipython==9.9.0
ipython_pygments_lexers==1.1.1
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
librt==0.7.7
lightning-utilities==0.15.2
line_profiler==5.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mpmath==1.3.0
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
packaging==25.0
parso==0.8.5
pathspec==1.0.1
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytokens==0.3.0
pytorch-lightning==2.6.0
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
safetensors==0.7.0
scipy==1.16.3
sentry-sdk==2.48.0
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.5.12
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
stringzilla==4.6.0
sympy==1.14.0
text-unidecode==1.3
timm==1.0.24
torch==2.9.1
torchinfo==1.8.0
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.23.1
wcwidth==0.2.14
wheel==0.45.1
yarl==1.22.0
